src/test/java/uk/gov/hmcts/reform/profilesync/domain/UserProfileSyncExceptionTest.java
package uk.gov.hmcts.reform.profilesync.domain;

import static org.assertj.core.api.Assertions.assertThat;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.junit.MockitoJUnitRunner;
import org.springframework.http.HttpStatus;


@RunWith(MockitoJUnitRunner.class)
public class UserProfileSyncExceptionTest {

    @Test
    public void should_create_exception_correctly() {
        String message = "this-is-a-test-message";
        UserProfileSyncException exception = new UserProfileSyncException(HttpStatus.NOT_FOUND, message);

        assertThat(exception).hasMessage(message);
        assertThat(exception).isInstanceOf(RuntimeException.class);
        assertThat(exception.getHttpStatus()).isEqualTo(HttpStatus.NOT_FOUND);
        assertThat(exception.getErrorMessage()).isEqualTo("this-is-a-test-message");
    }
}
